Base architecture for abstraction of the Maps Javascript API implementation has been loosely established. The core abstraction is based on 3 particular elements: the html file, the bound object, and the browser object. All are dependent on one another to produce a working implementation.
IGoogleMapsBrowser is an interface which mainly enforces two specific functionality:  ExecuteScriptAsync() and BindObject(); It acts as a browser specific to the maps functionality, but mainly it enforced those the two aforementioned methods. Other members are enforced by the interface but are less important. ExecuteScriptAsync and BindObject can both be implemented by an IMapBoundObject.

A map bound object serves as a mechanism for maintaining state and altering state of the map - through the browser object. Rather, the browser object does not directly execute high-level API tasks such as (AddMarker). While a map bound object and browser are used in tandem, all high level tasks should be executed through a map bound object.

Lastly, the javascript template executes scripts on the web side. It would be nice to allow a less strict enforcement of the JS scripts to use, and allow a JS implementation to be provided and also be compliant with some contract, like IMapBoundObject - but this is difficult. The template works for now.

Finally, I will expect major challenges with concurrency / race conditions. Locks? Hmm.