Latest progress consisted of considering how the MapBoundObject is going to abstract the idea of "BoundElements" while maintaing state with the api/map. I was trying to force the design to use my wrapper's element types (GoogleMapElement > Marker) to keep things consistent across the project, but I need a readonly DTO to map back and fourth to the browser.
Secondly, I learned just how simple it is to pass JSON to JS (I mean.... JS...ON). So, the data in transit will in fact be JSON. This will be a much better solution than passing positional arguments.

A BoundMarker record (DTO) is being toyed with currently. The next step was determining how to keep uniqueness in the collection of added elements to the map. While Google's API supports elements with same properties, I need a way to keep track of a specific object. Or do I?

This unique enforcement led me down the road of HashSets and building a little LockSafeHashList<TItem,THash> to store unique items. Is there not a built-in for this? Maybe im just a fool but I had time to kill.

Project is messy but one day it will be g.r.e.a.t.